refactor: put fields.go into a new jsonpath package
